requirements.txt
# Основные зависимости для каршеринг приложения

# Flask и веб-сервер
Flask==2.3.3
Flask-CORS==4.0.0

# База данных MySQL
mysql-connector-python==8.1.0

# Дополнительные библиотеки для Flask
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Для работы с JSON и HTTP запросами
requests==2.31.0
urllib3==2.0.7

# Для работы с датой и временем
python-dateutil==2.8.2

# Библиотеки для GUI (админ-панель)
# Встроенные в Python: tkinter (не требует установки)

# Дополнительные утилиты
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4

# Для разработки (опционально)
# python-dotenv==1.0.0

# Системные библиотеки
six==1.16.0